Added images.Need to fix bug

+++ OnlineKart/src/com/shopping/controller/DispatcherServlet.java
@@ -84,11 +84,11 @@ public class DispatcherServlet extends HttpServlet {
 		else if (userPath.equals("/home")) {
 			ProductService productService = new ProductService();
 			List<Product> productsList = productService.getAllProducts();
-			HashMap<Integer,List<String>> imagesMap = new HashMap<Integer,List<String>>();
+			HashMap<Long,List<String>> imagesMap = new HashMap<Long,List<String>>();
 			// Iterate through the list of products and retrieve the Image URL's
 			for (Product product : productsList) {
 				List<String> imagesList = productService.getImageURL(product.getProductId());				
-				imagesMap.put(product.getProductId(), imagesList);
+				imagesMap.put((long) product.getProductId(), imagesList);
 			}
 			getServletContext().setAttribute("productsList", productsList);
 			getServletContext().setAttribute("images", imagesMap);
